requirements.txt
Flask==3.1.3
flask-cors==6.0.2
pandas==3.0.1
pyarrow==23.0.1
numpy==2.4.3
python-dotenv==1.2.2
google-genai
faiss-cpu